+++ tools/src/main/java/it/eng/tools/service/ApplicationPropertiesService.java
@@ -1,6 +1,7 @@
 package it.eng.tools.service;
 
 import java.util.HashMap;
+import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
@@ -19,13 +20,13 @@ import it.eng.tools.exception.ApplicationPropertyErrorException;
 import it.eng.tools.exception.ApplicationPropertyNotFoundAPIException;
 import it.eng.tools.model.ApplicationProperty;
 import it.eng.tools.repository.ApplicationPropertiesRepository;
-import lombok.extern.java.Log;
+import lombok.extern.slf4j.Slf4j;
 
 /**
  * The PropertiesService class provides methods to interact with properties, including saving, retrieving, and deleting properties.
  */
 @Service
-@Log
+@Slf4j
 public class ApplicationPropertiesService {
 
 	private static final String STORED_APPLICATION_PROPERTIES = "storedApplicationProperties";
@@ -61,7 +62,7 @@ public class ApplicationPropertiesService {
 		List<ApplicationProperty> allProperties = null;
 
 		if(!StringUtils.isBlank(key_prefix)) {
-			allProperties = repository.findByKeyStartingWith(key_prefix, sortByIdAsc());
+			allProperties = repository.findByKeyStartsWith(key_prefix, sortByIdAsc());
 		} else {
 			allProperties = repository.findAll(sortByIdAsc());
 		}
@@ -81,8 +82,9 @@ public class ApplicationPropertiesService {
 	public Optional<ApplicationProperty> getPropertyByKey(String key) {
 		Optional<ApplicationProperty> propertyByMongo = repository.findById(key);
 		if(propertyByMongo.isEmpty()) {
-			log.warning(key + " not found in the db, try in application.properties");
-			//Try to keep value from applicatio.properties
+			log.warn(key + " not found in the db, try in application.properties");
+			//Try to keep value from application.properties
+			// TODO - Should we copy properties from env/property files to Mongo?
 			String propertyValueByApplicationProperty = env.getProperty(key);
 
 			if(propertyValueByApplicationProperty != null) {
@@ -254,6 +256,20 @@ public class ApplicationPropertiesService {
 	 */
 	public String get(String key) {
 		return env.getProperty(key);
+	}
+	
+	// called from ApplicationPropertiesConfiguration.init() - if we decide to go with env properties
+	public void copyApplicationPropertiesToEnvironment(Environment environment) {
+		try {
+			List<ApplicationProperty> allApplicationPropertiesOnMongo = getProperties(null);
+
+			for (Iterator<ApplicationProperty> iterator = allApplicationPropertiesOnMongo.iterator(); iterator.hasNext();) {
+				ApplicationProperty applicationProperty = (ApplicationProperty) iterator.next();
+				addPropertyOnEnv(applicationProperty.getKey(), applicationProperty.getValue(), environment);
+			}
+		} catch (ApplicationPropertyErrorException e) {
+			log.warn("Any property found in MongoDB!");
+		}
 	}
 
 }