+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 
+ - Removed daps properties from property file and using only from Mongo
+ - Removed protocol.authentication.enabled from property file and using it from Mongo
+ 
+## [0.1.1] - 21-02-2025
+
+### Changed
+
  - Removed catalog data from test scope initial_data.json
  - All IT inserts data before using it/verify logic and does cleanup after test
  - Force buffer flush for get external data and transfer file